+++ src/configurationmanager.ts
@@ -1,4 +1,4 @@
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
 
 type ConfigurationUpdateHandler = () => void;
 
@@ -17,6 +17,7 @@ export let refreshInterval: number = 10;
 export let showInternal: boolean = false;
 export let showAlpha: boolean = false;
 export let showBeta: boolean = false;
+export let showDeprecated: boolean = false;
 export let reportingShowInStatusBar: boolean = false;
 
 const debugConfiguration = () => {
@@ -25,33 +26,38 @@ const debugConfiguration = () => {
   console.log(`Show internal: ${showInternal}`);
   console.log(`Refresh alpha: ${showAlpha}`);
   console.log(`Refresh beta: ${showBeta}`);
+  console.log(`Refresh deprecated: ${showDeprecated}`);
   console.log(`Show in status bar: ${reportingShowInStatusBar}`);
 };
 
 export const updateConfiguration = () => {
   refreshInterval = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('control.refreshInterval') as number;
+    .getConfiguration("AlphaBETA")
+    .get("control.refreshInterval") as number;
 
   showAnnotations = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('control.showAnnotations') as boolean;
+    .getConfiguration("AlphaBETA")
+    .get("control.showAnnotations") as boolean;
 
   showInternal = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('phase.showInternal') as boolean;
+    .getConfiguration("AlphaBETA")
+    .get("phase.showInternal") as boolean;
 
   showAlpha = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('phase.showAlpha') as boolean;
+    .getConfiguration("AlphaBETA")
+    .get("phase.showAlpha") as boolean;
 
   showBeta = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('phase.showBeta') as boolean;
+    .getConfiguration("AlphaBETA")
+    .get("phase.showBeta") as boolean;
+
+  showDeprecated = vscode.workspace
+    .getConfiguration("AlphaBETA")
+    .get("phase.showDeprecated") as boolean;
 
   reportingShowInStatusBar = vscode.workspace
-    .getConfiguration('AlphaBETA')
-    .get('reporting.showInStatusBar') as boolean;
+    .getConfiguration("AlphaBETA")
+    .get("reporting.showInStatusBar") as boolean;
 
   onConfigurationUpdate.listeners.forEach((listener) => {
     listener();