+++ src/decorations.ts
@@ -1,4 +1,4 @@
-import { window, ThemeColor } from 'vscode';
+import { window } from 'vscode';
 
 export const internalDecorationType = window.createTextEditorDecorationType({
   backgroundColor: { id: 'AlphaBETA.internal' },
@@ -16,4 +16,10 @@ export const alphaDecorationType = window.createTextEditorDecorationType({
   backgroundColor: { id: 'AlphaBETA.alpha' },
   borderWidth: '2px',
   borderRadius: '4px',
+});
+
+export const deprecatedDecorationType = window.createTextEditorDecorationType({
+  backgroundColor: { id: "AlphaBETA.deprecated" },
+  borderWidth: "2px",
+  borderRadius: "4px",
 });